Move numberOfOpenItemsInString() from note.js to NPnote.js
